move ajv assertions
