+++ test/assertions/index.test.ts
@@ -2,7 +2,7 @@ import dedent from 'dedent';
 import * as fs from 'fs';
 import { createRequire } from 'node:module';
 import * as path from 'path';
-import { createAjv, runAssertion, runAssertions } from '../../src/assertions';
+import { runAssertion, runAssertions } from '../../src/assertions';
 import { validateXml, containsXml } from '../../src/assertions/xml';
 import { fetchWithRetries } from '../../src/fetch';
 import {
@@ -97,35 +97,6 @@ jest.mock('../../src/matchers', () => {
 });
 
 const Grader = new TestGrader();
-
-describe('createAjv', () => {
-  beforeAll(() => {
-    delete process.env.PROMPTFOO_DISABLE_AJV_STRICT_MODE;
-    jest.resetModules();
-  });
-
-  afterAll(() => {
-    delete process.env.PROMPTFOO_DISABLE_AJV_STRICT_MODE;
-  });
-
-  it('should create an Ajv instance with default options', () => {
-    const ajv = createAjv();
-    expect(ajv).toBeDefined();
-    expect(ajv.opts.strictSchema).toBe(true);
-  });
-
-  it('should disable strict mode when PROMPTFOO_DISABLE_AJV_STRICT_MODE is set', () => {
-    process.env.PROMPTFOO_DISABLE_AJV_STRICT_MODE = 'true';
-    const ajv = createAjv();
-    expect(ajv.opts.strictSchema).toBe(false);
-  });
-
-  it('should add formats to the Ajv instance', () => {
-    const ajv = createAjv();
-    expect(ajv.formats).toBeDefined();
-    expect(Object.keys(ajv.formats)).not.toHaveLength(0);
-  });
-});
 
 describe('runAssertions', () => {
   const test: AtomicTestCase = {
